# models/posted.go
package models

import "time"

//内存对齐
type Post struct {
	ID          int64     `json:"id,string" db:"post_id"`
	AuthorID    int64     `json:"author_id" db:"author_id"`                          // 作者id
	CommunityID int64     `json:"community_id" db:"community_id" binding:"required"` //社区id
	Status      int64     `json:"status" db:"status"`                                //状态
	Title       string    `json:"title" db:"title" binding:"required"`               //标题
	Content     string    `json:"content" db:"content" binding:"required"`           //内容
	CreateTime  time.Time `json:"create_time" db:"create_time"`                      //生成时间
}

// ApiPostDetail 帖子详情接口的结构体
type ApiPostDetail struct {
	AuthorName       string             `json:"author_name"` //作者
	VoteNum          int64              `json:"vote_num"`    //投票分数
	*Post                               //嵌入帖子结构体
	*CommunityDetail `json:"community"` //嵌入社区信息
}